Remove Ninja from CI
This change removes an external dependency from our CI flow, that has
previously shown to be unreliable and not always available. This should
help to unbreak the build.

Ninja is used in LLVM/MLIR builds to speed up the build process,
as it is often "faster" than make. However, Ninja provides a
large benefit mostly in situations where everything is already
built and only a small number of files are changed. In these
situations, it often drops the build time from 10s of seconds to
below one second. In our CI runs, the build takes about 1h as
all of LLVM/MLIR must be rebuilt. A reduction of 10s of seconds
has here negligible effects. Hence, the use of 'make' over 'Ninja'
should not negatively affect CI performance.