add doc generation and push to separate repo
To start, just put the documentation in the index.md file
in the documentation repo, as we only
have one documentation file so far.
Once we start expanding our documentation to multiple files, we can
think of adding a more suitable structure to the site, and also to move
the action to a separate workflow.
Please find the documentation site here:
https://rodonisi.github.io/llhd-docs/