Use an experimental MLIR dialect symbol
This removes the need for the last MLIR patch. We currently only need
MLIR patches to register the LLHD dialect symbol.  MLIR provides a
couple of experimental dialect symbol tags which we can use without the
need to patch MLIR. This patch moves us to one of these experimental
tags, as this simplifies the installation of LLHD for new users. In the
future, we can always move back to use the LLHD symbol tag if we find
the current configuration does not work for some reason.

I also incremented the key for the github action cache to make sure the
cached MLIR version does not have an LLHD symbol, which is necessary to
ensure our CI verifies correctly that everything builds even without
this symbol.