add allowUnregisteredDialects flags in llhdc
The commit bab5bcf8fd60c9df3d1517e5ca22109c785016d1 introduced a new argument in the MlirOptMain method for allowing unregistered dialects.
For this reason I added a new flag to the llhdc tool, which defaults to false. The flag has to be added to tests where generic operations, that are not part of the dialect, are used.